Add test code from researching and experimenting
This is some experimental code for calibrating a camera and then detecting the position and orientation of ArUco markers. Several static images are taken beforehand and put in the "calibrationImgs/" directory so that the camera can be calibrated before the video feed starts. The "charucoboard.png" image in "arucoMarkers/" was printed out and used for the calibration images. The "arena.png" and "arena_with_robots.png" images in "arucoMarkers/" were printed out and used for experimenting with detection after calibration.