Fix rotation average calculation error
The average of all line rotations was being thrown off by some lines being made of points in the opposite order.

Also, now being able to compare the two different methods has shown that rotation of an average line instead of average of rotation of all lines seems to be more accurate and consistent.