Divided screen into quadrants for rotation
This uses horizontal and vertical axis to divide the screen into four quadrants to get a rotation between 0.0 & 360.0

This also moves the line generation for a robot's markers into the Robot::make_lines() function

This also adds experimentation with having all sides of the marker being used for line creation instead of just one pair of opposing sides. How good/bad this is comparatively is inconsistent so far. Sometimes its better, sometimes its worse, sometimes it doesn't seem to make much of a difference.