Add camera, detector classes; rework state vars
This reworks how GlobalState handles information updates. All of the mutex locking has been encapsulated within a class. StateVariables also now has an atomic version rather than just an "update available" flag

This adds the start of an abstract base Camera class to be extended upon for connecting to and controlling cameras

This adds the start of several detector classes that will be used in conjunction with a camera to detect different types of objects

This reworks the CMakeLists.txt to use FindProtobuf instead of manually adding the protobuf dependencies