Update CMakeLists for MSVC
This moves stdc++fs link library command to only execute on non-MSVC targets since it isn't needed and doesn't work with MSVC

This also removes "designated initializers" for structs from the codebase (from markerdetector.cpp specifically). Designated initializers aren't compatible with MSVC unless "/stdc++latest" is specified, but doing so breaks ASIO because of a missing `inline` statement in `use_awaitable.hpp`. The fix for this doesn't seem to be present yet in the latest release of ASIO