Add Spinnaker SDK to CMakeLists.txt
This adds the Spinnaker SDK to the CMake dependencies requirement. Given the root installation path for the SDK, the include directory will be added, 32bit or 64bit libraries will be added depending on target platform, and release or debug versions of the library will be added depending on compilation target

This also adds Spinnaker SDK to the README dependencies list

The SDK installation path must be set with the cmake variable "SPINNAKER_SDK". Usage example: -DSPINNAKER_SDK=C:\Dev\SpinnakerSdk