Add support for opencv device id connections
This adds the ability to connect to a device through opencv with a device id, i.e. '0'. This just checks if the given url string is actually an integer and, if it is, converts it to an int and uses that to connect. Otherwise it just connects with the given string like normal